@@ -16,36 +16,74 @@
 #include <string>
 #include <map>
 
+
 // Typedefs for tag hierarchy
-using NodeType =
-    std::map<std::string, boost::variant<bool,
-    int,
-    double,
-    pybind11::none,
-    std::string,
-    std::vector<int>,
-    std::vector<std::string>>>;
+using Variants = boost::variant<
+        bool,
+        int,
+        double,
+        pybind11::none,
+        std::string,
+        std::vector<int>,
+        std::vector<std::string>
+>;
 
+using NodeType = std::map<std::string, Variants>;
 
+// How to print a NodeType
+static inline std::ostream& operator << (std::ostream& os, const NodeType& node)
+{
+    os << std::endl << "    {" << std::endl;
+    for (const auto& [key, value]: node) {
+        if (value.type() == typeid(std::string)) {
+            os << "        " << key << " : " << boost::get<std::string>(value) << std::endl;
+        }
+    }
+    os << "    }" << std::endl;
+    return os;
+}
+
+// How to print an std::vector of NodeTypes
+static inline std::ostream& operator << (std::ostream& os, const std::vector<NodeType>& nodes)
+{
+    os << std::endl << "[" ;
+    for (auto const node: nodes) {
+        os << node;
+    }
+    os << "]" << std::endl;
+    return os;
+}
+
+#pragma GCC visibility push(default)
+/*
+ * Modelhierarchy is the struct which is associated with all vertices in the graph
+ */
 struct Modelhierarchy {
     std::string id;
     NodeType properties;
 };
 
+/*
+ * Connection is the struct which is associated with all edges in the graph
+ */
 struct Connection {
     int id;
 };
+#pragma visibility pop
 
 using EdgeLabelMap = std::map<std::string, int>;
 
 using TagHierarchyGraph = boost::adjacency_list<
-        boost::listS, boost::vecS, boost::bidirectionalS,
+        boost::listS, boost::listS, boost::bidirectionalS,
         Modelhierarchy, Connection>;
 
 using TagHierarchyT = boost::labeled_graph<TagHierarchyGraph, std::string>;
 
 using VertexT = boost::graph_traits<TagHierarchyGraph>::vertex_descriptor;
+using VertexIter = boost::graph_traits<TagHierarchyGraph>::vertex_iterator;
 using EdgeT = boost::graph_traits<TagHierarchyGraph>::edge_descriptor;
+
+using VertexDescMap = std::map<VertexT, size_t>;
 
 namespace boost {
 namespace serialization {
@@ -92,3 +130,4 @@ static std::map<ApplicationMode, const char*> AppModeName = {
     {ApplicationMode::STATE_MANAGER, "statemanager"},
     {ApplicationMode::SERVER, "server"}
 };
+
