@@ -9,16 +9,16 @@
 #include <opencensus/trace/with_span.h>
 #include <opencensus/context/with_context.h>
 
-Command::Command() {
-
+Command::Command(std::string name) : name_(name) {
+    TagHierarchy::Register(*this, name);
 }
 
 std::vector<NodeType>
 Command::Process(std::vector<NodeType>& request) {
     static opencensus::trace::AlwaysSampler sampler;
     auto result = std::vector<NodeType>();
-    auto span = opencensus::trace::Span::StartSpan(CommandName(), nullptr, &sampler);
-    span.AddAnnotation("Start running command " + CommandName());
+    auto span = opencensus::trace::Span::StartSpan(name_, nullptr, &sampler);
+    span.AddAnnotation("Start running command " + name_);
     {
         opencensus::trace::WithSpan ws(span);
         try {
@@ -26,7 +26,7 @@ Command::Process(std::vector<NodeType>& request) {
             span.AddAnnotation("Finished running command");
             if (!result.empty() &&
                 result.at(0).count("error")) {
-                const auto log_string = std::string("The ") + CommandName() +
+                const auto log_string = std::string("The ") + name_ +
                                         std::string(" command failed with the error ") +
                                         boost::get<std::string>(result.at(0).at("error"));
                 span.AddAnnotation(log_string);
@@ -41,6 +41,10 @@ Command::Process(std::vector<NodeType>& request) {
     return result;
 }
 
+DispatchFunction Command::Function() {
+    return [this](std::vector<NodeType>& in) {return this->Process(in);};
+}
+
 TagHierarchyGraph &Command::GetGraph() {
     return TagHierarchy::GetGraph();
 }
@@ -51,8 +55,4 @@ std::map<std::string, VertexT> &Command::GetVertices() {
 
 VertexT &Command::GetRoot() {
     return TagHierarchy::GetRoot();
-}
-
-void Command::Register(Command &in) {
-    TagHierarchy::Register(in);
 }
