@@ -1,15 +1,11 @@
-//
-// Created by Petter Moe Kvalvaag on 2019-10-03.
-//
 #pragma once
 
 #include "tag_hierarchy/tag_hierarchy.h"
 #include "tag_hierarchy/commands/command.h"
 
-class Nodes : Command {
+class Delete : Command {
 public:
-    Nodes(std::string name);
-
+    explicit Delete(std::string name);
 private:
     std::vector<NodeType> ProcessRequest(std::vector<NodeType>& request) override;
 };
