@@ -1,25 +1,13 @@
-#define BOOST_TEST_MODULE tag_hierarchy
-#define BOOST_TEST_DYN_LINK
+#define BOOST_TEST_MODULE tag_hierarchy_test
+#define BOOST_TEST_MAIN
+#include <boost/test/unit_test.hpp>
+
 
 #include "models/models.h"
 
-#include <boost/test/unit_test.hpp>
-
 #include "tag_hierarchy/tag_hierarchy.h"
+#include "tag_hierarchy/unittests/fixture.h"
 
-struct Fixture {
-  Fixture() {
-     BOOST_TEST_MESSAGE( "Constructing test fixture" );
-     auto modelhierarchy = std::vector<NodeType>(
-     #include "hierarchy_dump.cpp"
-     );
-     TagHierarchy::Handle(modelhierarchy);
-  }
-
-  ~Fixture() {
-      BOOST_TEST_MESSAGE( "teardown fixture" );
-  }
-};
 
 BOOST_FIXTURE_TEST_SUITE( NodesTest, Fixture );
 BOOST_AUTO_TEST_CASE( test_nodes )
@@ -29,12 +17,10 @@ BOOST_AUTO_TEST_CASE( test_nodes )
           {std::string("parentId"), pybind11::none()}}}
     );
     auto response = TagHierarchy::Handle(query);
-    BOOST_TEST(boost::get<std::string>(response[0].at("name")) == "Level1-1");
-    BOOST_TEST(boost::get<std::string>(response[1].at("name")) == "Level1-2");
+    BOOST_TEST(boost::get<std::string>(response[0].at("name")) == "Level1-2");
+    BOOST_TEST(boost::get<std::string>(response[1].at("name")) == "Level1-1");
     BOOST_TEST(boost::get<int>(response[1].at("levelno")) == 1);
     BOOST_TEST(boost::get<pybind11::none>(response[1].at("parent_id")) == pybind11::none());
-    //Fixture::next_parent_id = boost::get<std::string>(response[1].at("id"));
-    // BOOST_TEST(boost::get<std::string>(response[1].at("is_modelelement")) == false);
 }
 
 BOOST_AUTO_TEST_CASE( test_l2_nodes ) {
@@ -56,26 +42,20 @@ BOOST_AUTO_TEST_CASE( test_l3_nodes ) {
           {std::string("parentId"), parent_id}}}
     );
     auto response = TagHierarchy::Handle(query);
-    BOOST_TEST(boost::get<std::string>(response[1].at("name")) == "Level1-2->Level2-1->Level3-2");
+    BOOST_TEST(boost::get<std::string>(response[1].at("name")) == "Level1-2->Level2-1->Level3-1");
     BOOST_TEST(boost::get<int>(response[0].at("levelno")) == 3);
     BOOST_TEST(boost::get<std::string>(response[1].at("parent_id")) == parent_id);
     std::cout << boost::get<std::string>(response[1].at("id"));
-    //BOOST_TEST(boost::get<bool>(response[0].at("is_modelelement")) == false;
 }
 
 BOOST_AUTO_TEST_CASE( test_l4_nodes ) {
-    const auto parent_id = std::string("0b8b675f-0afb-7772-6409-81a069aedb2a");
+    const auto parent_id = std::string("cfc07435-e890-9bf7-b4c8-645dced9ab73");
     auto query = std::vector<NodeType>(
         {{{std::string("command"), std::string("nodes")},
           {std::string("parentId"), parent_id}}}
     );
     auto response = TagHierarchy::Handle(query);
     BOOST_TEST(boost::get<int>(response[0].at("levelno")) == 4);
-    // TODO make support for model element folders
-    // assert response_l4.data[0]['type'] == NodeType.MODELELEMENTFOLDER
-    // assert response_l4.data[1]['is_modelelement'] is True
-    // assert response_l4.data[0]['parent_id'] == response_l3.data[1]['id']
-    // assert response_l4.data[1]['parent_id'] == response_l4.data[0]['id']
 }
 
 BOOST_AUTO_TEST_SUITE_END()
